FIX: temorarly switched pnpmode to lose
Fix the loading of the `@apollo/gateway` package.
It tries to load `node-fetch` but does not list it as a dependency,
witch will result in an error in `@yarnpkg/pnpify`.
This errors is difficult to see as the loading fo this module is done
dynamically through nest witch hides it from us.

This gives a lot of anoying warnigns though ):.